parsers/isobmff: return `undefined` in number-returning ISOBMFF parsing functions
Previously, ISOBMFF parsing function (e.g. one parsing the Track
Fragment Decode Time: the first presentation time available in the
segment) were returning `-1` for any invalid case (e.g. the
corresponding box was not found).

The base behind this code was already here when I came working on this
project but I suppose that it has to do with that code being low-level
and as such, copying C idiomatic way of doing that.

Returning `null` or `undefined` instead will allow to leverage
TypeScript's type checker to ensure those case are handled and it's
also much more idiomatic.

It may have (MAY) performance repercussions but we're talking about a
tiny amount of code running in a highly optimized JIT compiler most of
the time so who really knows.