demo: integrate RxPlayer directly in the demo bundle
This commit tries to simplify the building phase of the demo, especially
now that we'll be more and more dependant on using external RxPlayer
tools in it.

Previously, the demo code was separated into two big files:
  - the RxPlayer "lib" file as `lib.js`
  - the demo file as `bundle.js`

Both of them were then included in the demo's `index.html` and
`bundle.js` relied on the RxPlayer being accessible through
`window.RxPlayer`.

This worked perfectly but there are multiple drawbacks here:

  - Going through `window.RxPlayer` corresponds to the legacy way of
    depending on the RxPlayer. We advise people to rely on usual ES6
    `import` statements everywhere else.

  - In this mode, we cannot rely on features such as tree-shaking.

  - In this mode, we cannot access to the RxPlayer's tools as they are
    not exposed through `window` but only when imported through
    `"rx-player/tools"`. This is becoming even more important now that
    the `VideoThumbnailLoader` and the `StringUtils` tools are coming,
    as both could become useful to the demo.

What I decided here is to update the `webpack-demo.config.js` file which
was previously only used to compile the demo down to a `bundle.js` to
now also be able to build the RxPlayer's code (by adding typescript
rules and by setting compile-time constants there).

I then relied on a webpack features called "alias" which allows to
easily gives alias to modules, so as the RxPlayer defined in
`src/index.ts` can be always imported through "rx-player" and, more
importantly, tools can be imported through "rx-player/tools".

The build time seems to not be impacted by this change which is a good
news. Another good news is that the whole demo size was slightly
reduced.

A downside is that the re-building happening after a file modification
seems to take a little more time, but it stays at an acceptable level (3
seconds on my PC for a modification on the RxPlayer's
`RepresentationStream`).