demo: begin converting demo files to typescript
This is just a proof-of-concept and proposal to convert our demo files
from JavaScript to TypeScript.

Both can here be used at the same time.

Note that this is just a proof of concept, not a true direction I want
to take necessarily.

There are very few advantages to transition to TypeScript in the demo.
The main one I see is developper comfort due to the improved
auto-completion and the type checking.
It may also be the occasion to eat our own dogfood by relying more on
our exported TypeScript APIs (exported type definitions).

The downside of switching the demo to TypeScript may be that outside
developpers not being pro-efficient in TypeScript might find its code
harder to follow.

Some people may also read the demo today to see an example of how they
could use the RxPlayer. By switching it to TypeScript, it may be harder
for them to read our code.