add "conformance" tests
This is a first try at the implementation of "conformance tests" which
allows to quickly tests browser API support on various environments
(devices and browser).

I added the `conformance` sub-directory in `tests`, which for now
contains two HTML files:

  - `media_key_system_access.html`: to quickly check which CDM is
    available on the platform and which configuration is accepted.

  - `hdcp_policy_check.html`: to quickly check if the
    `MediaKeys.getStatusForPolicy` is defined and to check if it can
    be used to detect in advance which minimum HDCP version is
    supported on the device.

The goal for now with those files is that they can be used as templates
to very quicly check and update to check the support of various features
without the need to run and deploy the whole RxPlayer package.

I chose to only define two plain HTML files for now just so it stays
very simple to update and deploy but I could imagine factorizing some
code in the future and adding a building step when/if we add a lot of
different tests there.

How each test can be configured and used is defined as a comment at the
top of the corresponding HTML file.