stream: avoid error when updating the priority of a synchronous request
This is a fix for a rare new bug we seem to encounter starting from our
`v3.21.1` release.

Note that I did not succeed to reproduce it, so this is mostly an
hypothesis.

The error might appear when updating the priority of a previously
non-pending segment request. If it's priority is high enough for it to
be launched right away and if this segment request ends synchronously
after being started (which is a relatively rare occurence, but it can be
the case for smooth initialization segments - as they are
locally-generated - or any locally (in-JS) cached initialization
segments), it is theorically possible that the `currentSegmentRequest`
variable we use in the `RepresentationStream` to refer to the... current
segment request is set to `null` directly (as it should because the
request just ended).

The corresponding logic does not expect the request to be finished
just synchronously after updating its priority, and thus did not
re-check whether `currentSegmentRequest` was still set.

A simple but still valid and functional fix is to just mutate
`currentSegmentRequest` just before actually doing the update.
The order here should not really matter (it might even be better
that way if re-entrancy become involved) and it ensures that the
possible side-effects are performed last.