segment_buffers: refacto `source_buffers` code into `segment_buffers`
> Overview

This is a complete refacto of the `source-buffers`-related code, that
was before split up in two directories:
   - `src/custom_source_buffers`: for "custom SourceBuffer"
     implementations

   - `src/core/source_buffers`: for the interface to those native and
     custom SourceBuffers.

Now, all that code has been merged in a single directory,
`src/core/segment_buffers` and the notion of "custom" or "native"
SourceBuffers have been erased.
We just implement `SegmentBuffers` now: media buffers which might be
based on a native SourceBuffer object, or not.

>> Reasons for this

There are multiple reasons for doing that refacto, the main ones being
the reasons defined in #763:
  - The `QueuedSourceBuffer` abstraction had no real utility for
    "custom" SourceBuffers, as those perform data insertion/removal
synchronously and thus do not need to be queued.
  - It gaves us some issues with low-latency contents

This PR do not include any of new feature.
It is only the first step for making further improvements on that part
of the code easier.

> What I've done

Despite the huge diff, I did few actual changes to the original logic.

  1. The main difference is that now, instead of needing each type of
     buffer (e.g. "audio", "video", "text", "image") to implement its
own native or custom SourceBuffer, they now need to implement the
higher-level `ISegmentBuffer` interface instead, which is roughly the
same interface than what was previously called the `QueuedSourceBuffer`.

  2. The `QueuedSourceBuffer` class is now only used for audio and video
     contents (the only types which actually need operations to be
queued) and as such is now called the `AudioVideoSegmentBuffer`.

  3. Previous custom SourceBuffers implementation are now wrapped in a
     more complex class implementing the `ISegmentBuffer` interface
(which again is based on the `QueuedSourceBuffer` API).
      Here, I had to repeat some code, especially around the
`SegmentInventory` between all implementations. We might want to create
an abstract class / a mixin to reduce code repetition in the future

  4. The huge majority of the diff was to just rename every occurence of
     the word `SourceBuffer` by the right term:
         - `SegmentBuffer`: is the name of the media buffer
           implementation for a given type and is roughly equivalent to
what the `QueuedSourceBuffer` was. Consequently, the
`SourceBuffersStore` has been renamed to `SegmentBuffers`
         - the expression `media buffers` is now used to refer to those
           buffers in general, without needing to use RxPlayer-specific
language
         - `SourceBuffer` is still used but should now refer only to the
           native `SourceBuffer` object.

   5. It is now possible to set the `inventoryInfos` when pushing a
      segment to `null` so it is not added to the inventory.
       This is for example useful when pushing init segment or in any
other tool which might want to use an `ISegmentBuffer` directly without
needing anything inventory-related

   6. The task of calculating the appendWindows (with the
      `APPEND_WINDOW_SECURITIES`) has been moved from the
`QueuedSourceBuffer` to the `RepresentationStream` to avoid needing to
do that operation each time.

   7. The creation of the native SourceBuffer is now done in the
      `AudioVideoSourceBuffer` instead of in the `SourceBuffersStore`
(because I could, but this can change)