segment_buffers: replace `ISegmentBuffer` interface by `SegmentBuffer` abstract class
This commit removes a lot of repetitions in the code happening because
multiple SegmentBuffer implementations all used the same code,
especially the code relative to the `SegmentInventory`.

What I did here is replace the `ISegmentBuffer` interface by a
`SegmentBuffer` abstract class with methods having a default
implementation.

This should do the same thing with only reduced code.