eme: add more EME-related logs and raise-up verbositiy for INFO level
During recent debugging sessions related to DRM, we found out that
important steps, such as the communication of a licence, were only
logged when the `LOGGER_LEVEL` is set to DEBUG. As this is a very
important step, it seems that log could be displayed even when the
level is set to INFO.

In that commit, I chose to display the important EME steps (creating the
MediaKeySystemAccess, creating the MediaKeys, attaching the MediaKeys,
creating the session, receiving and sending data from the CDM and the
possible clean-up on stop) in the INFO logger level (which might raise
up the verbosity on that level, but mostly at initialization).

Because the return of some of those steps is almost as important as the
action, I chose to also log when those succeed.

This commit also add logs for the DEBUG level at relatively important
steps that were not logged before, such as when the EMEManager asks for
the permission to attach a new MediaKeys instance.