segment_buffers: better check last init segment in the AudioVideoSegmentBuffer
In the AudioVideoSegmentBuffer, we check that the initialization segment
linked to a segment we want to push is the last initialization pushed before
pushing that segment.

To do that, we before only checked that initialization segment by reference.
This has the following issues:

  1. If a new initialization segment was made by mutating a previous
     one, we would have the same reference for different data. Here we
     would thus not push a segment that needs to be pushed.

  2. If we ever cloned the initialization segment's data, we would push
     it again unnecessarily. This should not lead to any bug but is not
     necessary.

This commit fixes the number 2 problem, without yet resolving the number
1 one for performance reasons (there's no need to go through the step of
  checking the content of an initialization segment each time for now).

The number 1 should be very easy to fix in the future with the current
modification if we ever want to do that though.