manifest: remove `transport` property and improve doc
The `Manifest` class is supposed to be totally agnostic of the transport
(e.g DASH, HSS etc.) used. As such, the `transport` property should be
unneeded.

For now, the only real usage I can think off would be to add information
to log and errors, but this is generally an already known property.