manifest/parsers: allows an `absoluteMinimum` position for live contents
Previously, we always considered the minimum position for a live content
to behave in one of two ways:

  1. stay constant

  2. raise linearly as time goes by and old segments are being removed
     from the CDN.

It appears however that we can also encounter a third, hybrid case:
For live contents that just begun, the real minimum reachable position
might be much higher than what the timeshift depth could make us
believe (because fewer segments have been generated until now).

So in that case, what we actually want to do is to consider both that
minimum reachable position (useful before the timeshift depth has been
reached) and the timeshift depth (which will be useful as soon as enough
segments to fill it  have been generated) to calculate the minimum
position.

I chose here to change the way the parsers declare their time bounds by
defining an object with three properties:

  1. `absoluteMinimum`: the absolute minimum time reachable when the
     manifest has been loaded. Note that for static contents, this value
     will most likely be the one we always rely on to calculate the
     minimum position.

  2. `timeshiftDepth`: the depth of the timeshift window.
     If it indicates a higher position than `absoluteMinimum` when we
     substract it from the calculated maximum position, we consider that
     result instead.

  3. `maximumTimeData`: An object allowing to calculate at any time the
     current maximum position. It has roughly the same form of the
     previous `maximumPosition` and serves the same purpose.