text/ttml: allows multiple cues to be displayed at the same time
Previously, the RxPlayer didn't consider cases where multiple cues, each
with their own temporalities, could be displayed at overlapping times.
It only chose one single cue for a given time (which might contain
multiple subtitles).

However, having multiple cues at overlapping times happen to be a
frequent occurence when dealing with closed captions.

The gist of this commit is not very complicated: functions and methods
that were previously given a single HTML element are now given an array
of HTML elements.

However, just displaying the multiple elements as is on the screen led
to positionning issues when mutiple cues were displayed at the same
time: the first cue "pushed" the other cues below it.

Setting the corresponding cues' position to `absolute` instead of
`relative` appears to fix the issue.
I did not see any side-effects in test contents after doing that but I'm
as far as possible from an expert in CSS.

So I may just as well have broken everything there!

Good luck with the review!