stream: refacto Adaptation switch strategy
I spotted a possible issue with Adaptation switching which could led to
some "freeze" we see with multi-Period contents.

Basically, the `getAdaptationSwitchStrategy`, which is also called for
the first chosen Adaptation, would clean most of the time the buffer.

This "cleaning" task was actually performed by removing the data
corresponding to the time interval linked to the concerned Period, minus
some data around the current position to avoid decoding issues.

This behavior could lead to an issue for multi-Period contents, as it
could remove some segment's data from the end of the previous Period.
It could also remove some segment's data from the start of the next
Period.

This update:
  1. clean the buffer less often (we need to have another Adaptation
     from the same Period)
  2. apply paddings at the start and end if we play close to those