segment_buffers: add `codec` property to SegmentBuffers
Some logic in the RxPlayer might rely on the codec currently decoded by
the SegmentBuffers (for example for the audio and video SegmentBuffers,
where we might need to reload the MediaSource when the codec is totally
different, like in the audioTrackSwitchingMode feature).

This commit adds a `codec` property to SegmentBuffers which can be set
to a string or `undefined`. It is for now set to undefined (or really,
not set at all) for SegmentBuffers where it doesn't matter (text and
image) but to the codec of the SourceBuffer or - if it exists - of the
last pushed segment for an audio or video SegmentBuffer.

Note that this property already existed in previous versions but was
removed (because not useful until now) in the whole SegmentBuffer
refacto.