segment_buffers: Push init data again when changeType is called
This commit fix a not-yet released regression where an initialization
segment would not be pushed again - because it had the same content than
the last one pushed in that SourceBuffer - despite the fact that we
called the `changeType` API in the meantime.

This is a highly specific issue that only showed in tests for now.

To reproduce it, we need to have multiple Representations with the exact
same initialization segment data but with different codecs or profiles in
the Manifest. We also need to be on a browser that support the
`SourceBuffer.changeType` API.

In that very very rare situation, the RxPlayer would call the
`changeType` API when pushing segment data from a Representation with a
different codec (or profile) than the previous one. From my tests on
Firefox and Chrome, it seems that both expect to have the initialization
segment to be pushed again once that API is called.

Because the last pushed initialization segment data is exactly the same
than the one from the previous Representation (which has been pushed),
we do not push it again as an optimization. This led to a
`BUFFER_APPEND_ERROR`.

After multiple try to fix that, I opted for regrouping every operations
we do before an `appendBuffer` (`changeType`, updating
`timestampOffset`, setting the append windows etc.) in a
`_preparePushOperation` private method and do it only one time per
"Push" queue item in the `AudioVideoSegmentBuffer` (instead of one per
task, which would mean 1 or 2 per "Push" queue item, depending on if the
initialization segment has to be pushed), just before doing the first
`_appendBuffer` for that item.