avoid re-downloading a segment when it ends a lot before it should
This commit fixes a bug where the RxPlayer would request the same
segment in a loop if it really ended a lot after what we thought it
ended at. This usually happens only for the last segment in the content,
for various reasons and was easily reproducible by playing close to the
end of our multi-Period content in the demo.

As usual with requests happening in a loop, this is due to the RxPlayer
believing that segments that have been pushed have since been garbage
collected.

The source of the issue was in the (too-)complex SegmentInventory code.
Here, we link to each segments multiple properties including what
interest us here:

  - `end`: what the end of the segment should be once pushed (based on
    Manifest information, ISOBMFF metadata, or even on what has really
    been buffered after the segment has been pushed - if it's not too
    off).

  - `bufferedEnd`: the apparent end of the segment in the buffer
    currently. In contrast to `end` which is "what the end should be",
    `bufferedEnd` is: "what the end currently is".
    A difference between the two allows to know when a segment has been
    partially garbage-collected.

  - `precizeEnd`: this is a headache-inducing property, but it is
    basically a boolean set to `true` when the `SegmentInventory` has high
    confidence that the `end` property is really close to what should be
    the real end of the segment in the buffer when pushed.

    It is usually set to `true` after we compare where it should ends at
    with what the buffered ranges look like, when we see that both are
    close enough. Once `precizeEnd` is set to `true`, we won't try to
    update `end` anymore.

It was the way the three of them where updated in the
`guessBufferedEndFromRangeEnd` function that provoked the issue.

This function is called when we want to set those properties with the
end of a SourceBuffer's buffered range (when the segment linked to those
properties is suspected to be the last one of that range).

A subtle bug made it possible, after this function is called two times
for the same segment (which happens as this function is called
regularly), to update `end` to match the `end` of the current range -
which in this case is at a much higher value - but `bufferedEnd` to the
previous value of `end` (so a much lower value than the new `end`).

This can make the RxPlayer believe that the segment is garbage
collected, as `bufferedEnd` - or what seems to be the current place of
the segment in the buffer - becomes much lower than the value of `end` -
what should be its end.

This is all a little complicated but basically the issue is only when
the difference between the real end of the segment and the actual end is
superior to the `MAX_MANIFEST_BUFFERED_START_END_DIFFERENCE` config
property.
Here we would before this commit:
  1. update `bufferedEnd` to match `end`
  2. then, only if the segment looks like it is currently complete (not
     garbage collected), update `end` so it is equal to the range end
     and `precizeEnd` to `true`

Here, the second step looks off as there is no real reason why we would
update `end` to the range's end but not `bufferedEnd`.
Consequently I removed the step 2 entirely, nor `end` nor `precizeEnd`
is now updated in that case.

But this was not enough, as in the next clock tick, we would now enter
another case. If `bufferedEnd` is known and `precizeEnd` is set to
`false` (which is now the case) and the segment looks it is currently
complete, we would again update end to the range's end and `precizeEnd`
to true.

For this second case, I now added the condition that the difference
between the range end and the segment's end should be inferior or equal
to `MAX_MANIFEST_BUFFERED_START_END_DIFFERENCE`.

This is very absurdly complicated but after a lot of thinking, it seems
to make sense. Maybe the duplication of the whole
`MAX_MANIFEST_BUFFERED_START_END_DIFFERENCE` operation at several places
is not ideal though.

---

To simplify the code, I also did an unrelated modification.
`lastDeletedSegmentInfos` (see code) can now be `null` if no segment has
been deleted. I found it easier to understant that setting an initial
nonsensical value.