compat/eme: wait a small delay after loading session in some conditions
During our test on some set-top boxes, we found out a race condition in
some old versions of Chromium/Chrome/CEF where the `keyStatuses`
property was not properly polutated right after we get the answer from a
`mediaKeySession.load` call.

This made us believe that the MediaKeySession that was persisted did not
yet have any key stored, which is today a legitimate case, and
consequently the RxPlayer would close that MediaKeySession and re-create
a new one instead.

We found out that waiting a small delay, even 0 millisecond (through the
`setTimeout` native function), was enough to have the `keyStatuses`
property right in that case (Note: scheduling a microtask was not enough
however).

The resolution proposed in this commit is that, everytime the
`keyStatuses` property is empty, we should call `setTimeout(0)` before
communicating the MediaKeySession.

That resolution has the disadvantage of adding a small delay (as this `0`
millisecond translate sometimes to multiple real milliseconds) for
MediaKeySessions which have legitimately an empty `keyStatuses` property
(and thus should be closed then re-created).

As discussed, we might want to put that behind an option because of
that situation. But I'm still unsure about it.
The corresponding set-top-boxes might not be the only concerned by that
issue and I'm wondering if we didn't encounter that case on Chrome
Desktop some years back when we were first testing offline playback.
By running that code in every cases, we can fix the issue on every such
targets without needing to perform debug sessions and/or advertise about
the option in the already very complicated EME documentation. The only
downside being a fairly small cost of performance (some milliseconds for
a rare case).

We could also improve on that in the future by making sure we store
information on persistent MediaKeySession only when `keyStatuses` begin
to have its first entry.
That way, `keyStatuses` would very rarely be empty and the delay would
not be incurred on functional targets.