Seek back a consistent amount of time when switching track/bitrate
In some cases after switching the current track or bitrate, the RxPlayer
could be led to go into the `"RELOADING"` state, which corresponds to
visually a black screen (with nothing audible) before restarting
playback.

We could want to seek back some milliseconds when doing that.
For example, when switching the current audio track, it might make
sense to restart some time before, so the beginning of the sentence can
be heard again with the new track.

Before, we were seeking back at the last position emitted by the clock,
which can be anywhere between the current position and 1 second before.

This commit allows to seek back some hundreds of milliseconds configured
in `src/config.ts`. That delay depends on whether the track or the
bitrate is switched:
  - in case the bitrate is switched, we switch back 100 milliseconds
  - in case the audio track is switched, we switch back 700 milliseconds
  - in case the video track is switched, we switch back 100 milliseconds

I chose different values for audio and video due to both how they could
functionally be used (switching audio track needing generally to re-hear
what was being said, switching video track - to see another camera for
example - needs less context) as well as result of some tests I did on
my side:

  - For multiple audio tracks, I tested on:
    https://dash.akamaized.net/dash264/TestCasesIOP41/MultiTrack/alternative_content/6/manifest_alternative_lang.mpd

  - For multiple video tracks:
    https://dash.akamaized.net/dash264/TestCasesIOP41/MultiTrack/associative_content/1/manifest_associated_content_live.mpd

Both are listed as DASH-IF test vectors.

We could make them configurable values in the future, because I can see
that making sense functionally (to find out if it could be needed).

I'm still unsure on whether this delay should also be applied in the
future when we will perform a `"direct"` audio track switch without
`"RELOADING"`, but just by "flushing" the buffers.
It could make sense to stay consistent but that type of interruption is
not of the same scale: we re-begin to play much faster.

To discuss for this last point.