api: add `onCodecSwitch` option to loadVideo to reload on codec switch
At the time of v3.0.0, proper multi-codec handling was not integrated
and as such we never really asked ourselves the question of what to do
when we go from an audio and/or video in a codec A to an audio and/or
video in a _incompatible_ codec B.

Thankfully, the non-standard yet `SourceBuffer.changeType` API got
implemented at least in Chrome and Firefox, allowing to seamlessly
switch between codecs on the same SourceBuffer.

But this API is not implemented everywhere and even on targets where it
is, its effectiveness seems to vary.

As such, we still have to find a solution for those other targets so
they can play multi-codecs content reliably without hacks.

The obvious solution would be to just create a new SourceBuffer on those
targets, which requires to open a new MediaSource (what is called in
RxPlayer linguo as a "reloading").

But we might not want to do it in every cases, as like written earlier,
codec switching can work fine on some targets.
To have an efficient codec switching strategy on all targets, we are
thus left with the following choice:

  1. only reload based on the presence of specific APIs such as
     `changeType` (and browser detection?)

  2. always reload on codec switch, even on targets implementing the
     `changeType` API

  3. set an option allowing to choose between always reloading, never
     reloading or perhaps (not done yet) reloading only when the
     `changeType` API is unavailable.

This commit implements the third choice. I also want to implement an
"auto" mode in the future, which would be equivalent to the first
choice here, but I need to check first on some targets that have those
APIs, if `changeType` seams reliable enough.