eme/api: stop sending two times eme-related events
EME-related events like `warning` and `decipherabilityUpdate` were sent
two times in a row because we forgot to ignore the items emitted by one
of the Observable subscribed in `initialize_media_source.ts`.

As another Observable in that file already emitted EME-related events to
the API, each of those would be sent twice.

Thankfully that's the only bug here. The EMEManager being properly
shared, the real EME APIs are only called once and only the resulting
events emitted to the API (which is only used to trigger events to the
application) is sent twice.