manifest: `checkDiscontinuity` returns `null` instead of `-1`
The RepresentationIndex's `checkDiscontinuity` method returned `-1` when
no discontinuity was found at the given time.

This type of C-inspired values of giving negative numbers (even if some
JS native functions have the same thing like, indexOf) is not in the
style of the rest of the RxPlayer code.

It would make more sense, and most of all it would be type-checkable, to
use `null` instead` to mean: no discontinuity is found.

That way TypeScript can shout at us if we do not consider that
condition.