stream: change wording of events about "active" and "full" stream
The `RepresentationStream` previously sent ambiguous events such as
`"active-stream"` and `"full-stream"` which in effect were used to know
if the `RepresentationStream` was still downloading segments, to for
example allow a clever pre-loading of the next Period when the previous
one did not rely on network resources anymore

The notion of an "active" and "full" Stream is ambiguous as even a
"full" stream could still be pushing already-loaded segments.

I thus chose to rename the "active-stream" event to a
"downloading-segments" event and the "full-stream" event to a
`"download-finished"` event.