clock: re-define `IStalledStatus` interface so it does not include `null` (#855)
The previous type definition for `IStalledStatus`, which is used in the
clock to anounce that current playback is stalled and why, included the
`null` value, for when we had no active stall.

Including a `null` value to our types is somewhat unusual in our code, we
generally prefer defining an interface and then defining a sum type
grouping both (`Interface | null`) when needed.

Also, this commit does a second minor thing. The `getRange` function was
called two times in the clock: once implicitely in `getLeftSizeOfRange`
and then explicitely.
I re-wrote that code so we only need to calculate that value once per
clock tick.