manifest: a segment's timescale is now always set to `1`
The `ISegment` object had two time attributes, `time` and `duration`,
both expressed in a timescaled unit, indicated by `timescale`.

The main point of doing that was to always obtain round numbers, thus
making sure that we could retrieve the same value after doing additions
and substractions (to avoid bugs due to the `0.1 + 0.2 !== 0.3`
situation: https://0.30000000000000004.com/).

This timescale was generally the same one than the one taken from the
Manifest, as a Manifest generally has the same constraint.

However, in most area of the RxPlayer's code, we use seconds and
relatively rarely perform operations on those attributes.
When we do, it is furthermore concentrated in the very few parts of the
code which actually generate those attributes, which means we could be
careful by doing the operations always the same way without too much
effort.

Converting from timescale to seconds is easy but gives relatively ugly
code where timestamps in seconds are compared (see
`get_needed_segments.ts`). Plus the facts that we had some bugs in
the past where we forgot to perform the update, and that we perform
that conversion quite often in the code.

I hesitated on multiple improvements, like adding both the time in
seconds and the timescaled one to segments or even creating a `Segment`
class which would have a `getStart` method (which would return a
memoized property).
But this would have added a lot of redundancy which can lead to a whole
other class of bugs.

After discussing with Guillaume Renault, we decided that getting rid of
the old `time` and `duration` property to replace them by their
equivalent in seconds was perhaps the best solution - re-calculating
them always the same way when they are needed and making use of the
`privateInfos` property when we cannot be sure that the operation will
always happen the same way.

But there's a catch! `time` and `duration` were both defined in our API
documentation as being expressed in `timescale`. However, this can be
simply fixed by always setting the `timescale` property to `1` (and
perhaps deprecating that property? To discuss).