Detect and skip over most discontinuities
The previous behavior over discontinuities in the player was divided in
two strategies:

  1. If the discontinuity is small enough, just skip it when
     encountered.
     The "small enough" was actually "less than 1 second" which can mean
     that segments could be skipped, most notably between Periods - where
     those could be smaller - if they weren't loaded fast enough
     (technically, we also would have needed to load the first segment
     from the next Period, so this kind of problematic behavior would
     be rare enough).

  2. If the player is stalled, check that we have a discontinuity at
     that time in the Manifest. If we have, skip it.

The first has the enounced problem of false positives and the second had
more of an architectural problem: the `Stream` was listening to and
reacting to playback issues, a task normally given to modules which are
part of `Init` (`Stream` being mostly the module knowing which segments
have to be loaded and pushed).

More importantly, and the reason behind this commit, was that a category
of discontinuity was not handled: small discontinuities between Period
not anounced in the Manifest, when segments from the following Period
are not loaded until that Period is played.

This looks rare but it becomes more and more common with features
reloading the content such as `audioTrackSwitchingMode` or the
upcoming `onCodecSwitch`. In the current situation, the same problem can
be encountered when switching between two Periods when one of them does
not have either audio or video but the other has (thus needing to reload
when the other Period is played).

--

This commit attempts to propose a general strategy for any kind of
discontinuity.

The idea is that the `Stream`, which decides which segments to download
to fill the buffer, could anounce time intervals in the buffer for which
no segment seems to be available.

That data can then be used by the `Init` through a new `StallAvoider` by
correlating the player stalling to one of these discontinuities. When
detected, the `StallAvoider` would then seek over those.

--

The current code also brings a sensible change to the behavior of the
player: if a `startAt` that was given to the player indicated a position
not found in the Manifest (e.g. long before the first position), we
would before stop and throw a `MEDIA_STARTING_TIME_NOT_FOUND` error.

Now, the behavior is too automatically seek to the next Period, also
sending a `DISCONTINUITY_ENCOUNTERED` warning.

--

There are still some areas I'm not satisfied with, notably:

  - should there be a maximum seekable time when discontinuities are
    encountered (e.g. should a 60 min discontinuity be skipped
    automatically)?

  - should discontinuities anounced in the Manifest be handled
    differently than discontinuities detected in the buffer (but not in
    the Manifest)?

  - the current logic in the Stream which finds discontinuities in the
    buffer is run at every clock ticks. This means at minimum every
    seconds. This is not necessary, but I'm still unsure of if it can
    become a problem (we have arguably the same situation for the logic
    deciding which segments to download) and of what to do here.

  - the code doing discontinuity detection is very complicated to
    follow.
    There may be occasions to improve on that.