manifest: throw MANIFEST_PARSE_ERROR again on no supported codec
Previously, when a Manifest contained audio and video tracks but none
for a given type were supported (e.g. no supported audio tracks but
supported video tracks), we failed with a `MANIFEST_PARSE_ERROR` error.

Today in the same situation, the player goes on and just play the other
type of media (the one that is supported!).

This new behavior was not really wanted but was a side-effect of how we
now treat Representations in unsupported codecs. Where before we
filtered them out, now we keep them, but anounce that their codec is
not supported currently. This broke the logic throwing the
`MANIFEST_PARSE_ERROR` error.

Wanting one of the other behavior both make sense to me, but I think the
previous one may be the right default one: If a content contains both
audio and video, we generally want to be able to make both available.

This commit allows to retrieve the old behavior. I'm open to suggestions
of not returning to it though.