eme: create temporary MediaKeySession when persistent fails to update
License policies could disallow a license to be persisted. In that
case, what should be the RxPlayer's strategy when the user wanted the
RxPlayer to use persistent licenses?

There are two main possible answers here:
  - Either we throw, as the user's request cannot be met
  - Either we create a temporary MediaKeySession in those cases, while
    perhaps warning the user about it

The RxPlayer previously implicitely used the former solution, which
seems logical (at least to me) for most use cases.

But we now encounter new cases which makes the situation more complex:
some contents have multiple associated licenses. In a case where some
licenses can be persisted but some other cannot, what should we do here?

This commit tries to implement the second solution. When an "update"
fails for a persistent MediaKeySession (for any type of error as we
cannot detect the true reason why), we create a temporary
MediaKeySession instead, and re-perform the license request.