tests: add conformance tests relative to the initial seek
While debugging the rx-player on a video game console (it has the number
4 in its name, no, no 6 before, though that would be nice to also
support those) we noticed that the rx-player did not play any live
content.

But hey, it just works on PC, so we can just say that this is the
console fault and we're good!
Sadly no, they checkmated us by putting us in front of dash.js
which does play live contents on it.

After some investigations, we noticed that the main difference between
our player and dash.js here was the moment at which the seek to the
initial position is performed:

  1. The RxPlayer does it as soon as it can: when the "loadedmetadata"
     event is received.

  2. dash.js does it late but still soon enough to theorically make no
     difference: just after the first media segments have been pushed.

---

We noticed that under the first scenario, the `currentTime` on that game
console did not change for long, it was soon moved back to `0`.

Under the second however, no problem, the RxPlayer plays the content
without any issue.

I personally think that the first solution still has advantages over the
second (chances to see the first picture faster, media element
properties are directly coherent...) so I'm not yet ready to just
switch our strategy.

Instead, I added here two conformance tests, that we will bring to that
console's tech team:
  - `initial_seek_after_loaded_metadata.html` tests that the RxPlayer
    strategy works
  - `initial_seek_after_segment_push.html` tests that the dash.js
    strategy works

Adding those permanently as conformance tests will help us detect
similar issues on future target but also allow us to test the evolution
of this issue on that console.

I almost forgot, I had to write those ES5-style because that console
does not support ES6. So yeah, just `var` and XHRs everywhere like in
the good (bad) old days.